backend/requirements.txt
python-dotenv
requests
spotipy
youtube_search
ffmpeg
ffmprobe
yt_dlp